Refactors the SqlValue-related code
Updates `SqlValue` to include more variants for the various sizes of
integers and floats. The variants for SqlValue were also feature gated
so that the enum is only as big as it needs to be. Executors should
only enable the variants that they support. As a result of the feature
gating, the dependency on metafor had to be dropped. All of the code
related to the `SqlValue` type was also moved into its own module.